requirements.txt
feedparser==6.0.10
spacy==3.7.2
scikit-learn==1.3.2
google-generativeai==0.3.1
python-dotenv==1.0.0
flask==2.3.3
python app.py